requirements.txt
ffmpeg-python==0.2.0
requests~=2.31.0
pillow~=10.2.0
pytube~=15.0.0